forensic_toolkit/requirements.txt
pandas
chardet
beautifulsoup4
openpyxl
lxml  # optional but recommended
faster-whisper
ffmpeg-python
imageio-ffmpeg
rich
# Vision processing
torch
torchvision
transformers
ultralytics
pytesseract
opencv-python
packaging
# Email processing
extract_msg  # For MSG files
pypff        # For PST files (libpff bindings)
mailbox      # Built-in but listed for clarity
# OCR and image processing
pytesseract>=0.3.10
pdf2image>=1.16.3
opencv-python>=4.8.1.78
pillow>=10.0.0
numpy>=1.24.0

# FNB statement parsing
pdfplumber>=0.10.3